Add error handling and retrying
